+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2026-01-31
 
 ### Added
+- **Auto-Hide focus watcher**: New `auto_hide` option in the `[window]` block to automatically hide the window when it loses focus.
 - **Systray menu on Linux**: Add tray functionality via `ksni` to linux with full menu and shortcut display
 
 ### Fixed